Removed volumetric window mode on visionOS
